Add styles to body element to show Style Inheritance.
When a style is added to top element it is added to all its child elements.
Styles for body element is added to all the child text elements.
Styles are not applied to forms and buttons as they are special elements.
Styles like color, font-family are inherited while other styles may not.