Add alignment for text and inline elements under block element.
text-align : Align the text to the center of webpage.
For every block level element, centering its content can be done.
Text-align cannot be done for span elements as they donot have width for it.

Inline elements can be treated just like as text dislpayed in case of laying out.
So, text-align property can be used for block element containing bunch of inline elements within it.